# Shift TruffleBoundary around in Truffle::RegexOperations