Poking around in `Truffle::RegexpOperations` and tweaking the TruffleBoundary. It's slightly faster. 

## Changes

- Moved `regex.matcher` into a private function with Truffle Boundary
- Add boundary to `makeRegexForEncoding`
- Remove TruffleBoundary from `createMatcher`

## Before
```
truffleruby 21.2.0-dev-36afa76d, like ruby 2.7.3, GraalVM CE JVM [x86_64-darwin]
Warming up --------------------------------------
gsub! from .normalize_path
                       239.460k i/100ms
sub! from .normalize_path
                       354.894k i/100ms
          basic gsub    76.491k i/100ms
Calculating -------------------------------------
gsub! from .normalize_path
                          6.385M (± 8.9%) i/s -     31.609M in   5.010893s
sub! from .normalize_path
                          4.107M (± 4.0%) i/s -     20.584M in   5.020501s
          basic gsub      2.742M (± 8.7%) i/s -     13.615M in   5.018347s
```

## After
```
truffleruby 21.2.0-dev-36afa76d, like ruby 2.7.3, GraalVM CE JVM [x86_64-darwin]
Warming up --------------------------------------
gsub! from .normalize_path
                       159.989k i/100ms
sub! from .normalize_path
                       352.139k i/100ms
          basic gsub    74.733k i/100ms
Calculating -------------------------------------
gsub! from .normalize_path
                          6.851M (± 8.5%) i/s -     33.918M in   5.001933s
sub! from .normalize_path
                          4.328M (± 4.0%) i/s -     21.833M in   5.054152s
          basic gsub      2.812M (± 6.6%) i/s -     13.975M in   5.000743s
```

Using this benchmark: 

```ruby
# frozen_string_literal: true
require 'benchmark/ips'
puts RUBY_DESCRIPTION
path = +"foo/bar/baz"

Benchmark.ips do |x|
  x.report('gsub! from .normalize_path') do
    path.gsub!(/(%[a-f0-9]{2})/, &:upcase)
  end

  x.report('sub! from .normalize_path') do
    path.sub!(%r{/+\z}, "")
  end

  x.report('basic gsub') do
    'haystack'.gsub(/y/, 'X')
  end
end

```