@chrisseaton 2021-06-09
Was 50% faster on more specifically just the `gsub` wasn't it?